// interfaces/tokenInterface.ts
export interface TokenBalance {
    authorities?: any[];
    burnt?: boolean;
    compression?: Record<string, any>;
    content?: Record<string, any>;
    creators?: any[];
    grouping?: any[];
    id?: string;
    interface?: string;
    mutable?: boolean;
    ownership?: Record<string, any>;
    royalty?: Record<string, any>;
    supply?: Record<string, any>;
    token_info?: any;
}